+++ src/ResourceManager/Compute/Commands.Compute/Strategies/AsyncCmdletExtensions.cs
@@ -22,6 +22,9 @@ namespace Microsoft.Azure.Commands.Compute.Strategies
 {
     static class AsyncCmdletExtensions
     {
+        public static void WriteVerbose(this IAsyncCmdlet cmdlet, string message, params object[] p)
+            => cmdlet.WriteVerbose(string.Format(message, p));
+
         /// <summary>
         /// Note: the function must be called in the main PowerShell thread.
         /// </summary>
@@ -46,7 +49,7 @@ namespace Microsoft.Azure.Commands.Compute.Strategies
                             if (!taskProgress.IsDone)
                             {
                                 var config = taskProgress.Config;
-                                activeTasks.Add(config.Name + " " + config.Strategy.Type);
+                                activeTasks.Add(config.GetFullName());
                             }
                             progress += taskProgress.GetProgress();
                         }