+++ src/ResourceManager/Compute/Commands.Compute/Strategies/Network/NetworkSecurityGroupPolicy.cs
@@ -13,13 +13,13 @@
 // ----------------------------------------------------------------------------------
 
 using Microsoft.Azure.Management.Internal.Resources.Models;
-using Microsoft.Azure.Commands.Compute.Strategies.ResourceManager;
 using System.Linq;
 using Microsoft.Azure.Management.Internal.Network.Version2017_10_01.Models;
 using Microsoft.Azure.Management.Internal.Network.Version2017_10_01;
 using System;
 using Microsoft.Azure.Commands.Common.Strategies;
 using Microsoft.Azure.Management.Compute.Models;
+using Microsoft.Azure.Commands.Compute.Strategies.ComputeRp;
 
 namespace Microsoft.Azure.Commands.Compute.Strategies.Network
 {
@@ -27,7 +27,6 @@ namespace Microsoft.Azure.Commands.Compute.Strategies.Network
     {
         public static ResourceStrategy<NetworkSecurityGroup> Strategy { get; }
             = NetworkStrategy.Create(
-                type: "network security group",
                 provider: "networkSecurityGroups",
                 getOperations: client => client.NetworkSecurityGroups,
                 getAsync: (o, p) => o.GetAsync(
@@ -37,14 +36,14 @@ namespace Microsoft.Azure.Commands.Compute.Strategies.Network
                 createTime: _ => 15);
 
         public static ResourceConfig<NetworkSecurityGroup> CreateNetworkSecurityGroupConfig(
-            this ResourceConfig<ResourceGroup> resourceGroup, string name, Func<int[]> getOpenPorts)
+            this ResourceConfig<ResourceGroup> resourceGroup, string name, int[] openPorts)
             => Strategy.CreateResourceConfig(
                 resourceGroup: resourceGroup,
                 name: name,
                 createModel: _ => new NetworkSecurityGroup
                 {
-                    SecurityRules = getOpenPorts()
-                        .Select((port, index) => new SecurityRule
+                    SecurityRules = openPorts
+                        ?.Select((port, index) => new SecurityRule
                         {
                             Name = name + port,
                             Protocol = "Tcp",
@@ -63,10 +62,9 @@ namespace Microsoft.Azure.Commands.Compute.Strategies.Network
             this ResourceConfig<ResourceGroup> resourceGroup,
             string name,
             int[] openPorts,
-            Func<OperatingSystemTypes> getOsType)
+            ImageAndOsType imageAndOsType)
             => resourceGroup.CreateNetworkSecurityGroupConfig(
                 name,
-                () => openPorts ?? 
-                    (getOsType() == OperatingSystemTypes.Windows ? new[] { 3389, 5985 } : new[] { 22 }));
+                imageAndOsType.UpdatePorts(openPorts));
     }
 }