+++ src/ResourceManager/ServiceFabric/Commands.ServiceFabric.Test/ScenarioTests/TestServiceFabric.cs
@@ -12,7 +12,6 @@
 // limitations under the License.
 // ----------------------------------------------------------------------------------
 
-using Microsoft.Azure.Commands.Management.Storage.Test.ScenarioTests;
 using Microsoft.Azure.Commands.ServiceFabric.Commands;
 using Microsoft.Azure.ServiceManagemenet.Common.Models;
 using Microsoft.WindowsAzure.Commands.Common;
@@ -32,7 +31,7 @@ namespace Microsoft.Azure.Commands.ServiceFabric.Test.ScenarioTests
             AddAzureRmServiceFabricNodeType.dontRandom = true;
             ServiceFabricCmdletBase.WriteVerboseIntervalInSec = 3;
             ServiceFabricCmdletBase.RunningTest = true;
-            ServiceFabricCmdletBase.TestThumbprint = "2394f562bf05258059fe32c0d7c63024ead13096";
+            ServiceFabricCmdletBase.TestThumbprint = "2F51AC39C590551FC7391A7A0A187A67BF8256CA";
         }
 
         [Fact]
@@ -100,9 +99,37 @@ namespace Microsoft.Azure.Commands.ServiceFabric.Test.ScenarioTests
 
         [Fact]
         [Trait(Category.AcceptanceType, Category.CheckIn)]
+        public void TestNewAzureRmServiceFabricCluster()
+        {
+            TestController.NewInstance.RunPsTest("Test-NewAzureRmServiceFabricCluster");
+        }
+
+        [Fact]
+        [Trait(Category.AcceptanceType, Category.CheckIn)]
+        public void TestAddAzureRmServiceFabricNode()
+        {
+            TestController.NewInstance.RunPsTest("Test-AddAzureRmServiceFabricNode");
+        }
+
+        [Fact]
+        [Trait(Category.AcceptanceType, Category.CheckIn)]
+        public void TestRemoveAzureRmServiceFabricNode()
+        {
+            TestController.NewInstance.RunPsTest("Test-RemoveAzureRmServiceFabricNode");
+        }
+
+        [Fact]
+        [Trait(Category.AcceptanceType, Category.CheckIn)]
         public void TestAddAzureRmServiceFabricNodeType()
         {
             TestController.NewInstance.RunPsTest("Test-AddAzureRmServiceFabricNodeType");
+        }
+
+        [Fact]
+        [Trait(Category.AcceptanceType, Category.CheckIn)]
+        public void TestRemoveAzureRmServiceFabricNodeType()
+        {
+            TestController.NewInstance.RunPsTest("Test-RemoveAzureRmServiceFabricNodeType");
         }
     }
 }