+++ src/ResourceManager/Automation/ChangeLog.md
@@ -18,6 +18,12 @@
         - Additional information about change #1
 -->
 ## Current Release
+* Fix issue with Default Resource Group in CloudShell
+
+## Version 4.3.0
+* Fixed issue with importing aliases
+
+## Version 4.2.1
 * Added alias from New-AzureRmAutomationModule to Import-AzureRmAutomationModule
 
 ## Version 4.2.0