+++ src/ResourceManager/EventHub/ChangeLog.md
@@ -18,6 +18,9 @@
         - Additional information about change #1
 -->
 ## Current Release
+* Fix issue with Default Resource Group in CloudShell
+
+## Version 0.6.1
 * Fix bug in Get-AzureRmEventHubGeoDRConfiguration help
 
 ## Version 0.6.0