+++ src/ResourceManager/HDInsight/ChangeLog.md
@@ -18,6 +18,10 @@
         - Additional information about change #1
 -->
 ## Current Release
+* Fix issue with Default Resource Group in CloudShell
+
+## Version 4.1.0
+* Fixed issue with importing aliases
 
 ## Version 4.0.2
 * Added Location Completer to -Location parameters allowing tab completion through valid Locations