+++ src/ResourceManager/RecoveryServices.SiteRecovery/ChangeLog.md
@@ -1,4 +1,4 @@
-﻿<!--
+<!--
     Please leave this section at the top of the change log.
 
     Changes for the current release should go under the section titled "Current Release", and should adhere to the following format:
@@ -18,6 +18,10 @@
         - Additional information about change #1
 -->
 ## Current Release
+* Fixed issue with cleaning up scripts in build
+
+## Version 0.2.2
+* Fixed issue with importing aliases
 
 ## Version 0.2.1
 * Fixed assembly loading issue that caused some cmdlets to fail when executing